backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyMuPDF
requests
pydantic
pydantic-settings
openai
pinecone
